+++ df-downloader-ui/src/main.tsx
@@ -5,13 +5,15 @@ import { Provider } from "react-redux";
 import { BrowserRouter } from "react-router-dom";
 import "./index.css";
 import { store } from "./store/store.ts";
+import { SnackbarProvider } from "notistack";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <Provider store={store}>
-      <BrowserRouter>
-        <App />
-      </BrowserRouter>
+      <SnackbarProvider/>
+        <BrowserRouter>
+          <App />
+        </BrowserRouter>
     </Provider>
   </React.StrictMode>
 );