+++ df-downloader-ui/src/store/auth-user/auth-user.simple-actions.ts
@@ -1,0 +1,3 @@
+import { createAction } from "@reduxjs/toolkit";
+
+export const userLoggedOut = createAction("authUser/USER_LOGGED_OUT");