+++ df-downloader-ui/src/store/df-tasks/tasks.types.ts
@@ -2,7 +2,7 @@ import { TaskInfo, TaskPipelineInfo } from "df-downloader-common";
 import { QueryableState } from "../utils";
 import { DfUiError } from "../../utils/error";
 
-export interface DownloadQueueState extends QueryableState {
+export interface TasksState extends QueryableState {
   loading: boolean;
   taskPipelineIds: string[];
   taskPipelines: Record<string, TaskPipelineInfo>;