+++ df-downloader-ui/vite.config.ts
@@ -1,9 +1,14 @@
 import { defineConfig } from "vite";
 import react from "@vitejs/plugin-react-swc";
+import { nodePolyfills } from 'vite-plugin-node-polyfills'
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react()],
+  plugins: [react(),
+    nodePolyfills({
+      include: ['events'],
+    }),
+  ],
   envDir: "./environments",
   build: {
     rollupOptions: {